[WIP] use new sql interpolator everywhere
