+++ app/models/binary/Publication.scala
@@ -12,7 +12,7 @@ import play.api.libs.json.Format.GenericFormat
 import play.api.libs.json.{JsObject, Json}
 import slick.jdbc.PostgresProfile.api._
 import slick.lifted.Rep
-import utils.sql.{SQLClient, SQLDAO}
+import utils.sql.{SqlClient, SQLDAO}
 import utils.ObjectId
 
 import scala.concurrent.ExecutionContext
@@ -53,7 +53,7 @@ class PublicationService @Inject()(dataSetService: DataSetService,
   }
 }
 
-class PublicationDAO @Inject()(sqlClient: SQLClient)(implicit ec: ExecutionContext)
+class PublicationDAO @Inject()(sqlClient: SqlClient)(implicit ec: ExecutionContext)
     extends SQLDAO[Publication, PublicationsRow, Publications](sqlClient) {
   protected val collection = Publications
