+++ app/models/user/MultiUser.scala
@@ -12,7 +12,7 @@ import models.user.Theme.Theme
 import play.api.libs.json.Format.GenericFormat
 import play.api.libs.json.{JsObject, Json}
 import slick.lifted.Rep
-import utils.sql.{SQLClient, SQLDAO}
+import utils.sql.{SqlClient, SQLDAO}
 import utils.ObjectId
 
 import scala.concurrent.ExecutionContext
@@ -29,7 +29,7 @@ case class MultiUser(
     isDeleted: Boolean = false
 )
 
-class MultiUserDAO @Inject()(sqlClient: SQLClient)(implicit ec: ExecutionContext)
+class MultiUserDAO @Inject()(sqlClient: SqlClient)(implicit ec: ExecutionContext)
     extends SQLDAO[MultiUser, MultiusersRow, Multiusers](sqlClient) {
   protected val collection = Multiusers
