+++ app/models/user/time/TimeSpan.scala
@@ -6,7 +6,7 @@ import com.scalableminds.webknossos.schema.Tables._
 import play.api.libs.json.{JsValue, Json, OFormat}
 import slick.jdbc.PostgresProfile.api._
 import slick.lifted.Rep
-import utils.sql.{SQLClient, SQLDAO}
+import utils.sql.{SqlClient, SQLDAO}
 import utils.ObjectId
 
 import javax.inject.Inject
@@ -46,7 +46,7 @@ object TimeSpan {
 
 }
 
-class TimeSpanDAO @Inject()(sqlClient: SQLClient)(implicit ec: ExecutionContext)
+class TimeSpanDAO @Inject()(sqlClient: SqlClient)(implicit ec: ExecutionContext)
     extends SQLDAO[TimeSpan, TimespansRow, Timespans](sqlClient) {
   protected val collection = Timespans
